CMDIData now a generic interface. Old CMDIData is CMDIDataSolrImpl
