+++ vlo-importer/src/main/java/eu/clarin/cmdi/vlo/importer/CMDIData.java
@@ -1,31 +1,30 @@
+/*
+ * Copyright (C) 2018 CLARIN
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package eu.clarin.cmdi.vlo.importer;
 
-import java.util.Collection;
-
-import org.apache.solr.common.SolrInputDocument;
-import org.apache.solr.common.SolrInputField;
-
-import eu.clarin.cmdi.vlo.FieldKey;
-import eu.clarin.cmdi.vlo.config.FieldNameService;
 import eu.clarin.cmdi.vlo.importer.processor.ValueSet;
+import java.util.Collection;
+import java.util.List;
 
 /**
- * Represents a document of CMDI data.
+ *
+ * @author Twan Goosen <twan@clarin.eu>
  */
-public class CMDIData extends CMDIDataBaseImpl<SolrInputDocument> {
-
-    /**
-     * The associated solr document (not send to the solr server yet)
-     */
-    private SolrInputDocument doc;
-
-    public CMDIData(FieldNameService fieldNameService) {
-        super(fieldNameService);
-    }
-
-    public SolrInputDocument getDocument() {
-        return doc;
-    }
+public interface CMDIData<T> extends DocFieldContainer {
 
     /**
      * Sets a field in the doc to a certain value. Well, at least calls another
@@ -34,114 +33,66 @@ public class CMDIData extends CMDIDataBaseImpl<SolrInputDocument> {
      * @param valueSet
      * @param caseInsensitive
      */
-    public void addDocField(ValueSet valueSet, boolean caseInsensitive) {
-        addDocField(valueSet.getTargetFacetName(), valueSet.getValue(), caseInsensitive);
-    }
+    void addDocField(ValueSet valueSet, boolean caseInsensitive);
 
-    public void replaceDocField(ValueSet valueSet, boolean caseInsensitive) {
-        replaceDocField(valueSet.getTargetFacetName(), valueSet.getValue(), caseInsensitive);
-    }
+    void addDocField(String fieldName, String value, boolean caseInsensitive);
 
-    public void addDocField(String fieldName, String value, boolean caseInsensitive) {
-        if (fieldNameService.getFieldName(FieldKey.ID).equals(fieldName)) {
-            setId(value.trim());
-        } else {
-            handleDocField(fieldName, value, caseInsensitive);
-        }
-    }
-
-    public void replaceDocField(String name, String value, boolean caseInsensitive) {
-        if (this.doc != null) {
-            this.doc.removeField(name);
-        }
-        this.addDocField(name, value, caseInsensitive);
-    }
-
-    public void addDocFieldIfNull(ValueSet valueSet, boolean caseInsensitive) {
-        if (this.getDocField(valueSet.getTargetFacetName()) == null) {
-            this.addDocField(valueSet, caseInsensitive);
-        }
-    }
+    void addDocFieldIfNull(ValueSet valueSet, boolean caseInsensitive);
 
     /**
-     * Sets a field in the doc to a certain value. Before adding checks for
-     * duplicates.
+     * Add a meta data resource to the list of resources of that type.
      *
-     * @param name
-     * @param value
-     * @param caseInsensitive
+     * Whenever the type is not one of a type supported by the CMDI
+     * specification, a warning is logged.
+     *
+     * @param resource meta data resource
+     * @param type type of the resource
+     * @param mimeType mime type associated with the resource
      */
-    private void handleDocField(String name, String value, boolean caseInsensitive) {
-        if (doc == null) {
-            doc = new SolrInputDocument();
-        }
-        if (value != null && !value.trim().isEmpty()) {
-            if (caseInsensitive) {
-                value = value.toLowerCase();
-            }
-            Collection<Object> fieldValues = doc.getFieldValues(name);
-            if (fieldValues == null || !fieldValues.contains(value)) {
-                // if availability facet reduce tag to most restrictive
-                if (name.equals(fieldNameService.getFieldName(FieldKey.AVAILABILITY))
-                        || name.equals(fieldNameService.getFieldName(FieldKey.LICENSE_TYPE))) {
-                    reduceAvailability(name, value);
-                } else {
-                    doc.addField(name, value);
-                }
-            } // ignore double values don't add them
-        }
-    }
+    void addResource(String resource, String type, String mimeType);
 
-    @Override
-    public Collection<Object> getDocField(String name) {
-        if (doc == null) {
-            return null;
-        } else {
-            return doc.getFieldValues(name);
-        }
-    }
+    List<Resource> getDataResources();
+
+    Collection<Object> getDocField(String name);
+
+    T getDocument();
+
+    String getId();
 
     /**
-     * In case that Availability facet has more then one value use the most
-     * restrictive tag from PUB, ACA and RES. TODO: Move this to post processor
+     * Return the list of landing page resources.
      *
-     * @param field field to reduce availability values in
-     * @param value value to insert (add or replace)
+     * @return list of landing page resources
      */
-    protected void reduceAvailability(String field, String value) {
-        Collection<Object> currentValues = doc.getFieldValues(field);
-        // the first value
-        if (currentValues == null) {
-            doc.addField(field, value);
-            return;
-        }
-        int lvlNew = availabilityToLvl(value);
-        if (lvlNew == -1) {
-            // other tags, add them, uniqueness has already being checked
-            doc.addField(field, value);
-            return;
-        }
-        // current level of availability
-        int lvlCur = -1;
-        for (Object val : currentValues) {
-            int rhs = availabilityToLvl(val.toString());
-            if (lvlCur < rhs) {
-                lvlCur = rhs;
-            }
-        }
-        // if new values is more restrictive replace the old with new
-        if (lvlNew > lvlCur) {
-            SolrInputField fOld = doc.get(field);
-            SolrInputField fNew = new SolrInputField(field);
-            fNew.addValue(value); // new, more restrictive value
-            for (Object val : fOld.getValues()) {
-                // copy other tags
-                if (availabilityToLvl(val.toString()) == -1) {
-                    fNew.addValue(val);
-                }
-            }
-            doc.replace(field, fOld, fNew);
-        }
-    }
+    List<Resource> getLandingPageResources();
 
+    List<Resource> getMetadataResources();
+
+    /**
+     * Return the list of search page resources.
+     *
+     * @return the list
+     */
+    List<Resource> getSearchPageResources();
+
+    /**
+     * Returns list of all search interfaces (preferably CQL interfaces)
+     *
+     * @return list of search interface resources
+     */
+    List<Resource> getSearchResources();
+
+    /**
+     * Checks if any resources (metadata, landingpage, etc.) are available
+     *
+     * @return Returns true if at least one resource is available
+     */
+    boolean hasResources();
+
+    void replaceDocField(ValueSet valueSet, boolean caseInsensitive);
+
+    void replaceDocField(String name, String value, boolean caseInsensitive);
+
+    void setId(String id);
+    
 }
